fix: Remove cr hijack
I feel like a `<cr>` hijack is too much magic and unnecessary for my use case,
so I removed it.

Why don't just use the default `g:smartparis_hijack_enter = 0`? Well,
essentially, even with this option, it was conflicting with my autocomplete
mapping that completes on `<cr>`.

I could try and fix it, but I feel that the logic behind the way the original
author coded it might be a bit weird, and I don't even want to try understand
it.